surfsense_backend/pyproject.toml
[project]
name = "surf-new-backend"
version = "0.0.8"
description = "SurfSense Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "discord-py>=2.5.2",
    "docling>=2.15.0",
    "fastapi>=0.115.8",
    "fastapi-users[oauth,sqlalchemy]>=14.0.1",
    "firecrawl-py>=1.12.0",
    "github3.py==4.0.1",
    "google-api-python-client>=2.156.0",
    "google-auth-oauthlib>=1.2.1",
    "kokoro>=0.9.4",
    "langchain-community>=0.3.17",
    "langchain-unstructured>=0.1.6",
    "langgraph>=0.3.29",
    "linkup-sdk>=0.2.4",
    "llama-cloud-services>=0.6.25",
    "markdownify>=0.14.1",
    "notion-client>=2.3.0",
    "numpy>=1.24.0",
    "pgvector>=0.3.6",
    "playwright>=1.50.0",
    "jsonata-python>=0.6.0",
    "pypdf>=5.1.0",
    "python-ffmpeg>=2.0.12",
    "rerankers[flashrank]>=0.7.1",
    "sentence-transformers>=3.4.1",
    "slack-sdk>=3.34.0",
    "soundfile>=0.13.1",
    "spacy>=3.8.7",
    "en-core-web-sm@https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "static-ffmpeg>=2.13",
    "tavily-python>=0.3.2",
    "unstructured-client>=0.30.0",
    "unstructured[all-docs]>=0.16.25",
    "uvicorn[standard]>=0.34.0",
    "validators>=0.34.0",
    "youtube-transcript-api>=1.0.3",
    "litellm>=1.77.5",
    "langchain-litellm>=0.2.3",
    "elasticsearch>=9.1.1",
    "faster-whisper>=1.1.0",
    "celery[redis]>=5.5.3",
    "flower>=2.0.1",
    "redis>=5.2.1",
    "slowapi>=0.1.9",
    "chonkie[all]>=1.4.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "feedparser>=6.0.0",
    "bcrypt>=4.2.1,<5.0.0",
    "structlog>=24.1.0",
    "python-jose[cryptography]>=3.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.5",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "faker>=33.0.0",
    "pytest-mock>=3.14.0",
    "safety>=3.0.0",
    "bandit[toml]>=1.7.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

# Python 3.12
target-version = "py312"

[tool.ruff.lint]
select = [
    "E4",     # pycodestyle errors
    "E7",     # pycodestyle errors
    "E9",     # pycodestyle errors
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "C90",    # mccabe complexity
    "T20",    # flake8-print
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
]

extend-select = ["I"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes for strings.
quote-style = "double"

# Indent with spaces, rather than tabs.
indent-style = "space"

# Respect magic trailing commas.
skip-magic-trailing-comma = false

# Automatically detect the appropriate line ending.
line-ending = "auto"


[tool.ruff.lint.isort]
# Group imports by type
known-first-party = ["app"]
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.mccabe]
# Maximum cyclomatic complexity allowed
max-complexity = 10

[tool.pytest.ini_options]
# Minimum version of pytest
minversion = "8.0"

# Test discovery patterns
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Test paths
testpaths = ["tests"]

# Async support
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Coverage options
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-branch",
    "-v",
    "-ra",
]

# Markers for organizing tests
markers = [
    "unit: Unit tests that test individual functions/methods",
    "integration: Integration tests that test multiple components together",
    "e2e: End-to-end tests that test full workflows",
    "slow: Tests that take a long time to run",
    "security: Security-focused tests for vulnerabilities and attack vectors",
    "db: Tests that require database access",
    "auth: Tests related to authentication and authorization",
    "api: Tests for API endpoints",
    "services: Tests for service layer",
    "models: Tests for database models",
]

# Ignore warnings
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app", "app.*"]
